requirements.txt
numpy>=1.21.0
gurobipy>=9.5.0  # For MIP expert solver
pyyaml>=6.0  # For configuration files
scikit-learn>=1.0.0  # For additional ML utilities
matplotlib>=3.5.0  # For visualization
pandas>=1.3.0  # For data handling
